backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
chromadb==0.3.23
click==8.3.1
clickhouse-connect==0.10.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
duckdb==1.4.4
docx2txt==0.8
faiss-cpu==1.13.2
fastapi==0.128.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
hnswlib==0.8.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.3.4
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
lz4==4.4.5
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy-extensions==1.1.0
networkx==3.6.1
numpy==2.4.1
openai==2.16.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pandas==3.0.0
posthog==7.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic-core==2.41.5
pypdf==6.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==5.0.0
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid-utils==0.14.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
